refactor: note의 content 부분 폴더 구조 변경
