+++ src/app/(main)/note/[id]/_components/note-content/note-content.tsx
@@ -3,36 +3,12 @@
 import { useState, useRef } from 'react';
 import { css } from '@/../styled-system/css';
 import { ITextBlock } from '@/types/block-type';
-import PlusIcon from '@/icons/plus-icon';
-import GripVerticalIcon from '@/icons/grip-vertical-icon';
 import Block from './block';
+import BlockButton from './block-button';
 
 const blockContainer = css({
   display: 'flex',
   flexDirection: 'row',
-});
-
-const blockBtnContainer = css({
-  position: 'absolute',
-  left: '-3rem',
-  display: 'flex',
-  flexDirection: 'row',
-  pt: '0.2rem',
-});
-
-const blockBtn = css({
-  width: '1.5em',
-  height: '1.5rem',
-  padding: '0.2rem',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  borderRadius: '0.5rem',
-  cursor: 'pointer',
-
-  _hover: {
-    backgroundColor: '#F1F1F0',
-  },
 });
 
 const focusTextStyle = css({
@@ -69,6 +45,7 @@ const NoteContent = () => {
       ],
     },
   ]);
+
   const [isTyping, setIsTyping] = useState(false);
   const [isFocused, setIsFocused] = useState<boolean[]>([false]);
   const [isHover, setIsHover] = useState<boolean[]>([]);
@@ -96,16 +73,7 @@ const NoteContent = () => {
             onMouseEnter={() => handleMouseEnter(index)}
             onMouseLeave={() => handleMouseLeave(index)}
           >
-            {isHover[index] && (
-              <div className={blockBtnContainer}>
-                <div className={blockBtn}>
-                  <PlusIcon />
-                </div>
-                <div className={blockBtn}>
-                  <GripVerticalIcon />
-                </div>
-              </div>
-            )}
+            {isHover[index] && <BlockButton />}
             {isFocused[index] && block.children[0].content?.trim() === '' && (
               <div className={focusTextStyle}>
                 글을 작성하거나 AI를 사용하려면 '스페이스' 키를, 명령어를 사용하려면 '/' 키를
