+++ src/app/(main)/note/[id]/_components/note-header/line-info.tsx
