+++ src/constants/key-name.ts
@@ -1,0 +1,7 @@
+const keyName = {
+  enter: 'Enter',
+  backspace: 'Backspace',
+  slash: '/',
+};
+
+export default keyName;
